Update Comments.tmPreferences to preserve block indentation when commenting
This change allows the comment keyboard shortcut to be used to comment out lines of a recipe without requiring a manual fixup afterwards. Without this change, the pound sign `#` gets introduced at the start of the line, which confuses the Just parser if there are additional recipe lines after the commented-out line.